refactor(admin-service): TASK-BE-028b2 JWT cleanup, aspect single-path, canonical outbox envelope
- Drop OperatorRole enum; OperatorContext now carries operatorId + jti only
- OperatorAuthenticationFilter: token_type=admin claim (scope dropped),
  roles claim removed (permissions resolved via DB per rbac.md D5)
- Remove @PreAuthorize and in-use-case requireRole checks; single path
  via RequiresPermissionAspect @Order(200) with deny-by-default on
  controller mutations lacking @RequiresPermission
- AuditController cross-permission path now calls
  AdminActionAuditor.recordDenied(AUDIT_QUERY, "audit.read+security.event.read", ...)
  before throwing, yielding a DENIED admin_actions row + composite envelope
- AdminEventPublisher reshaped to canonical envelope:
  eventId / occurredAt / actor{type,id,sessionId=jti} / action{permission,endpoint,method}
  / target{type,id,displayHint} / outcome / reason
- New AdminPiiMaskingUtils (email/accountId/phone) feeds target.displayHint
- AdminActionAuditor.recordDenied: actionCode -> target_type mapping table;
  REQUIRES_NEW tx; emits canonical envelope for DENIED rows
- Flyway V0010: trg_admin_actions_finalize_only guards operator_id and
  permission_used; admin_operators.version BIGINT -> INT per spec
- data-model.md reconciled to started_at/completed_at (canonical); rbac.md
  recordDenied row mapping aligned
- AdminIntegrationTest seeds admin_operators + admin_operator_roles directly;
  OperatorJwtTestFixture issues canonical sub+jti+token_type=admin claims
- New tests: AdminPiiMaskingUtilsTest, AdminEventPublisherCanonicalEnvelopeTest,
  AuditDeniedCrossPermissionTest
- Remove every TODO(TASK-BE-028b) marker in apps/admin-service/

+++ apps/admin-service/src/main/java/com/example/admin/infrastructure/persistence/rbac/AdminOperatorJpaEntity.java
@@ -51,7 +51,8 @@ public class AdminOperatorJpaEntity {
     @Column(name = "updated_at", nullable = false)
     private Instant updatedAt;
 
+    // INT per data-model.md. Optimistic lock counter; monotonically increasing.
     @Version
     @Column(name = "version", nullable = false)
-    private long version;
+    private int version;
 }
